+++ cli/CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- `stacrs items` ([#360](https://github.com/stac-utils/stac-rs/pull/360))
+
 ### Fixed
 
 - **tokio** panic when validating ([#358](https://github.com/stac-utils/stac-rs/pull/358))